update api structure
